+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.1.0 - 2017-10-11
+* Added new module: removeNode
+* Added new module: sassGetVar
+* Added new module: sassHasVar
+
 # 2.0.1 - 2017-08-23
 * Fixed issue of not properly getting opposite values (getSides module)